requirements.txt
flask==3.0.2
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.24.0
Pillow>=10.0.0
requests>=2.31.0
transformers>=4.37.0
scikit-image>=0.22.0